OPS: Refactor Wallet header on TX list.
